Enemy Controller Refactor!
Enemies are now pretty standard RPG enemies. They have a couple of types - dumb will remain where they are. Guard will return to their starting position, patrolling is not implemented yet because I am lazy. They pass an action along on what to do when they arrive which isn't the most efficient but should make it so we can have the enemies hide when they are done moving