Morning Changes
There is now a functioning Title Screen. It is NOT pretty but it's ready to get painted
I have a healthbar tied to the player's health (when it's implemented)
There is an arrow always pointing at blub
Right-click to cancel a throw while aiming
Shadows are always on the ground under Zinnia and Blub
Code is in for giving the player Knockback. Right now it always goes off if you press Q, i'll eventually die this into the health system